+++ firmware/boron-ins-fsm/src/consoleFunctions.cpp
@@ -12,12 +12,11 @@
  *
  *
  */
-#include "consoleFunctions.h"
-
 #include "Particle.h"
+#include "consoleFunctions.h"
 #include "flashAddresses.h"
-#include "imDoorSensor.h"
 #include "stateMachine.h"
+#include "imDoorSensor.h"
 
 void setupConsoleFunctions() {
     // particle console function declarations, belongs in setup() as per docs
@@ -217,7 +216,10 @@ int ins_threshold_set(String input) {
 }
 
 // particle console function to get/set door sensor ID
-int im21_door_id_set(String command) {  // command is a long string with all the config values
+// command is a long string with all the config values
+int im21_door_id_set(String command) {
+    char buffer[64];
+    IMDoorID doorIDHolder;
 
     if (isValidIM21Id(command) == false) {
         return -1;
@@ -228,29 +230,33 @@ int im21_door_id_set(String command) {  // command is a long string with all the
 
     // if echo, publish current door ID
     if (*checkForEcho == 'e') {
-        IMDoorID holder;
-        EEPROM.get(ADDR_IM_DOORID, holder.byte1);
-        EEPROM.get((ADDR_IM_DOORID + 1), holder.byte2);
-        EEPROM.get((ADDR_IM_DOORID + 2), holder.byte3);
+        EEPROM.get(ADDR_IM_DOORID, doorIDHolder.byte1);
+        EEPROM.get((ADDR_IM_DOORID + 1), doorIDHolder.byte2);
+        EEPROM.get((ADDR_IM_DOORID + 2), doorIDHolder.byte3);
 
-        char buffer[64];
-        snprintf(buffer, sizeof(buffer), "{\"byte1\":\"%02X\", \"byte2\":\"%02X\", \"byte3\":\"%02X\"}", holder.byte1, holder.byte2, holder.byte3);
+        snprintf(buffer, sizeof(buffer), "{\"doorId\": \"%02X,%02X,%02X\"}", doorIDHolder.byte3, doorIDHolder.byte2, doorIDHolder.byte1);
         Particle.publish("Current Door Sensor ID: ", buffer, PRIVATE);
+
+        // put door ID in buffer for return value
+        snprintf(buffer, sizeof(buffer), "%02X%02X%02X", doorIDHolder.byte3, doorIDHolder.byte2, doorIDHolder.byte1);
     }
-    else  // else not echo, so we have a new door ID to parse
-    {
+    // else not echo, so we have a new door ID to parse
+    else {
         // parse input string and update global door ID
-        const char* byteholder1;
-        const char* byteholder2;
-        const char* byteholder3;
-        int split1 = command.indexOf(',');
-        byteholder1 = command.substring(0, split1).c_str();
-        globalDoorID.byte3 = (uint8_t)strtol(byteholder1, NULL, 16);
+        char* byteholder1;
+        char* byteholder2;
+        char* byteholder3;
+
+        int split1 = command.indexOf(',');                            // get index of first comma to delimit input
+        byteholder1 = strdup(command.substring(0, split1).c_str());   // get first byte of input and copy to holder variable
+        globalDoorID.byte3 = (uint8_t)strtol(byteholder1, NULL, 16);  // convert it to hex and set the third byte of the door ID
+
         int split2 = command.indexOf(',', split1 + 1);
-        byteholder2 = command.substring(split1 + 1, split2).c_str();
+        byteholder2 = strdup(command.substring(split1 + 1, split2).c_str());
         globalDoorID.byte2 = (uint8_t)strtol(byteholder2, NULL, 16);
+
         int split3 = command.indexOf(',', split2 + 1);
-        byteholder3 = command.substring(split2 + 1, split3).c_str();
+        byteholder3 = strdup(command.substring(split2 + 1, split3).c_str());
         globalDoorID.byte1 = (uint8_t)strtol(byteholder3, NULL, 16);
 
         // write new global door ID to flash
@@ -258,9 +264,13 @@ int im21_door_id_set(String command) {  // command is a long string with all the
         EEPROM.put((ADDR_IM_DOORID + 1), globalDoorID.byte2);
         EEPROM.put((ADDR_IM_DOORID + 2), globalDoorID.byte3);
 
+        // put door ID in buffer for return value
+        snprintf(buffer, sizeof(buffer), "%02X%02X%02X", globalDoorID.byte3, globalDoorID.byte2, globalDoorID.byte1);
+
     }  // end if-else
 
-    return 1;
+    // return door ID as int
+    return (int)strtol(buffer, NULL, 16);
 }
 
 int force_reset(String command) {